feat: allow to specify a default reference
Use the .docker/runx.yaml file and set the ref key

If some arguments are still passed, we are trying to see if that's an
image or not. If that's an image, use runx against it. But if we can't
access the image, use it as if it was an action

Signed-off-by: Yves Brissaud <yves.brissaud@gmail.com>