feat: allow to set local values
In a runx config manifest we can ask the user to fill any input. But
sometimes we want to be able to set it for a specific project.
We can now define it inside the local configuration .docker/runx.yaml.
The value from the local configuration will be used, and the user will
not be prompt for a value.
But if we want to temporaily override it, it's still possible using the
--ask flag.

Signed-off-by: Yves Brissaud <yves.brissaud@gmail.com>