feat: cache runx readme and config locally
So that we only need to perform a HEAD to get the digest of the
reference from the user. This HEAD request and the fact we are using the
digest as a key allow to invalidate cache each time the remote image
change.

Cache for a specific image with digest DIGEST contains these two items:

- ~/.docker/runx/cache/sha256/sha256:DIGEST/README.md
- ~/.docker/runx/cache/sha256/sha256:DIGEST/runx.yaml

Fixes #1

Signed-off-by: Yves Brissaud <yves.brissaud@gmail.com>