feat: support confirm (boolean) values
Use `optBool` to convert any value to a boolean to be used in the
template.

A new `type` field can be set for an option.
Possible values are:

- input: text field
- select: single select field
- confirm: ask for a yes/no confirmation

If no type is set, if some values are set type will be select, else it
will be an input. This is just for backward compatibility, type should
be set.

Signed-off-by: Yves Brissaud <yves.brissaud@gmail.com>